requirements.txt
numpy==1.24.3
pandas==2.0.3
matplotlib==3.7.2
seaborn==0.12.2
scikit-image==0.21.0
scikit-learn==1.3.0
tqdm==4.65.0
pyyaml==6.0.1
scipy==1.10.1
opencv-python==4.8.0.74
pillow==10.0.0
tensorflow==2.13.0
torch==2.0.1
torchvision==0.15.2
scenedetect==0.6.2
lpips==0.1.4
vmaf==2.3.1
